Revert "smith-waterman aln algo"
